requirements.txt
streamlit>=1.28.0
spacy>=3.7.0
pyttsx3>=2.90
pandas>=2.0.0
genanki>=0.13.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy<2.0,>=1.24.3
googletrans==4.0.0-rc1

